Package and serve aurutils via ichi-public repo
When the custom-repo role first executes, it will install aurutils by
downloading a PKGBUILD from the AUR and compiling it. It will then
create several custom repositories, one of which contains the aurutils
package.

On future runs of the custom-repo role, aurutils won't be installed in a
ports-like way, as it's already installed.

On future system updates, newer versions of aurutils will be installed
from the custom repository.