Make Python script better match PROMELA script
There is both a PROMELA model of Dijkstra's concurrent locking algorithm
and a Python script implementing the model. Both execute the algorithm
with some number of workers. Reduce the number of workers in the Python
script, so that it matches the number executed by the PROMELA script.
This has the additional advantage of making visual debugging easier, as
there are 25% fewer lines of output.