Switch from Jeepney and PySide2 to PyQt5
I'd like to write a couple simple little D-Bus apps, so that I can get a
more practical understanding of topics like designing and advertising an
interface (with XML files and all), receiving unicast messages, and
sending and receiving signals (multicast messages).

Unfortunately, Jeepney won't serve for this purpose. It's a delightful
little D-Bus library, but it can only be used to send unicast messages.
It can't be used to receive unicast messages, or send or receive
signals.

The most obvious solution is to use the Qt D-Bus toolkit for this
purpose, and to access it via PySide2. Unfortunately, PySide2 doesn't
support Qt D-Bus. And this is despite being the official Python bindings
for Qt!

PySide2 is a young library, and it is explicitly designed to mimic the
older third-party PyQt5 API. PyQt5's age shows here, as it provides
bindings for Qt D-Bus. This provides a nice path forward:

* Drop Jeepney in favor of PyQt5, as the latter supports the
  comprehensive Qt D-Bus toolkit.
* Drop PySide2 in favor of PyQt5, as it's silly to depend on both.

----

That should be the end of the story, but reality insistently imposed.
The Python bindings for Qt D-Bus make the underlying C++ interfaces
available in a very direct way. As an example of how weird things are,
the process of making an integer available as an unsigned int requires
something like the following:

    def get_qvariant_uint(val):
        obj = QVariant(val)
        obj.convert(QVariant.UInt)
        return obj

This could be OK if such techniques were well explained, but they
absolutely are *not* well explained.

This leaves the following options:

1. Power through. Learn the Qt application development paradigm, and
   figure out how to bridge the gap between it and Python.
2. Just use Qt for GUI stuff and find some other library for interfacing
   with D-Bus.
3. Switch to an entirely different toolkit.

The first option feels nauseating and is likely to be extremely
time-consuming.

The second option would be great if there actually were good Python
wrappers for D-Bus, but pydbus and txdbus appears unmaintained, and
jeepney and QtDBus are out of the running. The last Python wrapper is
PyGObject, which indirectly depends on GTK, and it would be weird for an
app to depend on both Qt for its GUI and GTK for its backend. This last
complaint holds even if the applicaiton is split into separate
components that communicate via an IPC mechanism (like D-Bus).

This leaves the third option. I'm not married to Qt, so this seems like
an option worth exploring.

----

The following web pages were especially useful when wrapping my head
around Qt D-Bus:

* https://codito.in/notifications-in-qt-over-dbus/
* https://techbase.kde.org/Development/Tutorials/D-Bus/Accessing_Interfaces