Migrate from PyQt5 to PyGObject
Using PyQt5 doesn't feel pythonic, and this can be dangerous to someone
who's use to writing Python code. For example:

* In certain circumstances, PyQt5 doesn't throw exceptions when errors
  are encountered, and instead returns error objects.
* When serializing Python data into a D-Bus message, PyQt5 guesses what
  types of data are being sent over the bus. For example, if the integer
  "5" is being sent over the bus, PyQt5 will guess whether you mean to
  send an INT64, UINT64, etc.

In addition, I found it very hard to discover conceptual information
about Qt's data types. What is a variant? How and when should I take a
Python variable and pack it into one of those data structures?

All in all, I found the experience to be unsatisfactory. This is the
motivation for trying out PyGObject. It, and the underlying libraries
like GTK and GIO, are much more comfortable to use, and I've found the
documentation much nicer.

The upshot of all this screwing around with toolkits is that I now have
a greater ability to experiment with D-Bus in a concrete way, and have
had a chance to think through some application design issues from top to
bottom.