Add unit tests for PyGObject GLib variants
GLib provides a "variant" data type, which can be used to store a data
type which isn't known until runtime. Variants are especially optimized
for data serialization, and therefore good for tasks like serializing
data to be sent over the dbus.

What happens if invalid data is passed to a variant, though? For
example, what if -1 or 2^16 is passed to a guint16 variant? This commit
adds unit tests which see what happens. These unit tests are written in
Python, not C, so they depend on PyGObject.