Jump from KDE to GNOME
GLib and GTK are quite nice. There's a strong possibility that I'll use
them when working on killjoy. I might as well spend some time with the
foremost DE which makes use of them, so that I'll have a better
intuitive understanding of the GNOME desktop application human interface
guidelines.