Package notification-generator
Add a desktop entry file and a PKGBUILD. This application should never
actually be shipped out to end users, as it's entirely educational.
Instead, my goal is to design, implement and ship a small D-Bus
application from star to finish, so that any glaringly obvious issues in
doing this sort of thing will be uncovered.

See:

* https://freedesktop.org/wiki/Specifications/desktop-entry-spec/
* https://www.archlinux.org/pacman/PKGBUILD.5.html