Create application to track extant units
Create a Python + PyGObject application that prints out the following:

* Unit disappearances.
* Unit appearances.
* All current units.

It has some issues as noted with comments in the code, it's *very*
lightly documented, and it has all the other issues one would expect
from a toy application. However, it does demonstrate one critically
important point: it is possible for one connection to subscribe to
multiple signals. This has positive implications for performance. It
means that one will *not* need to create one thread or process per
monitored unit. Instead, one thread or process can monitor all
interesting units, and dispatch the actual work of processing that
information to other threads or processes.