Drop HouseOfCuddles
Wireless routers currently advertise two APs: HouseOfCuddles and
HomeOfCuddles. The former is insecure, and the latter is secured with
WPA2. This division is in place with some intentions of doing cool
things like providing a throttled public wireless network.  However, I
don't have time to make this goal a reality, and this setup is more
complicated than a single-AP setup. The added complexity puts a damper
on my other networking goals, such as supporting the 5 GHz bands and
IPv6. Drop the insecure network, so that I can focus on other goals.