Add "media" group to webservers
Add `group-media` Ansible role, which creates a "media" group. Create
`/srv/media`, and configure it so that new files are automatically added
to the "media" group with rwX permissions. The purpose of this directory
is to hold files that are served by media servers such as subsonic and
jellyfin.

Add subsonic and ichimonji10 to the media group. Remove subsonic from
the ichimonji10 group.