Move completed torrents to /srv/media/torrents
Currently, incomplete torrents and complete torrents are both written to
`/var/lib/transmission/downloads`. This is workable, but there are two
frustrations:

1.  Incomplete downloads are visible in various media serving
    interfaces, including subsonic and the webserver file server.
2.  Users who wish to manage transmission's completed downloads must
    belong to the transmission group.

Simply changing the directory for completed downloads addresses the
former issue. Adding transmission to the "media" group and making use of
the permissions and ACLs set on /srv/media addresses the latter.