Rewrite btrfs-subvolume role
Remove the ability to customize how long snapshots are retained. This
functionality was neat, but not all that useful for my simple use case,
and it added a great deal of complexity to the role.

Let role install templates, like:

    btrfs-subvolume-delete@.service
    btrfs-subvolume-delete@.timer
    btrfs-subvolume-snapshot@.service
    btrfs-subvolume-snapshot@.timer

Do this instead of installing a custom-generated role per unit. This is
a far more maintainable approach, as there are no longer dozens of
nearly identical units installed.

Only snapshot `/mnt/btrfs/srv-media`, and drop snapshots of all other
subvolumes. This change reflects a change in filesystem layout and
organization.

Support the following relative layout:

    srv-media
    snapshots
        srv-media
            date-1
            date-2
            date-3