Package signal-desktop-bin instead of signal
Packaging signal 1.23.2 on a headless system causes a "cannot open
display" error to be raised. As per comments on the AUR [1], this may
occur when `electron -v` is executed.

Avoid the issue by packaging signal-desktop-bin instead. Also, update
the "workstation" role so that signal is uninstalled and
signal-desktop-bin is installed.

[1] https://aur.archlinux.org/packages/signal/