Implement VM networking with macvtap, not NAT
A downside of routing VM traffic is that IP subnets must be managed. If
one creates two virtual bridges, then one must assign a pair of IP
subnets to them, respectively. One must be careful when choosing which
IP subnets to assign to the virtual bridges. If one chooses a subnet
which conflicts with a subnet reachable via a physical interface, then
routing will break.

A downside of routing VM traffic with NAT is that you suffer all of the
problems inherent to NAT. Let's say that you set up a pair of VMs, where
each has SSH listening on port 22 and web servers listening on port 80
and 443. You'll need to set up port forwarding for six ports on the VM
host.

These two problems make it hard, on a practical level, to use VMs to
provide services. It should be as easy as possible to spin up a VM,
install a service, and configure the home router to push certain traffic
to that VM, and it's absolutely not.

Drop the NAT approach entirely. In the future, VMs will be connected to
the rest of the network with macvtap connections. This lets them act as
though they're bare-metal boxes on the same network link as the VM host.

For some information on macvtap, see:

* https://seravo.fi/2012/virtualized-bridged-networking-with-macvtap
* https://events.static.linuxfound.org/sites/events/files/slides/LinuxConJapan2014_makita_0.pdf
* https://virt.kernelnewbies.org/MacVTap