Add vim file type defs to vars files
Ansible host files lack a file extension. They have names such as "vars"
and "vault" instead of "vars.yml" or "vault.yml." As a result, text
editors can fail to perform syntax highlighting.

Add vim comments to the end of each "vars" file declaring it as a yaml
file, so that syntax highlighting works in vim.