Encrypt DNS queries
Stubby is a stub DNS resolver. It knows how to forward DNS queries to
upstream DNS resolvers, and how to do so either unencrypted or with TLS.
Do the following on all hosts:

1. Install, configure, start and enable stubby.
2. Configure dnsmasq to forward queries to stubby.

As far as I know, it is impossible to make dnsmasq merely prefer stubby.
A related thought is that it would be nice to configure stubby with
additional name servers.

Should dnsmasq need to fall back to regular DNS servers, it will still
enforce DNSSEC.