Refactor NN for fashion MNIST data
Keras is a toolkit that can, in many cases, abstract away the backend
toolkit being used. Backend toolkits include TensorFlow, CNTK and
Theano. Take advantage of this functionality: drop references to the
TensorFlow API, and reference the Keras API instead.

Also, change this import:

    import numpy as np

To this:

    import numpy

Many toolkits within the scientific and math community ask users to
import modules using abbreviations. For exmaple:

    import keras as K
    import matplotlib.pyplot as plt
    import numpy as np

This habit is presumptuous and harms legibility.