Manage aronia with Ansible
Currently, amelanchier works as a firewall, router, and wireless access
point. Unfortunately, the wireless access point functionality is
lacking. It's slow, spotty, and stuck on the 2.4 GHz channels.

Hardware-wise, aronia is a near-clone of amelanchier. The goal is to use
aronia as a testbed for new functionality. For example, when testing
alternate wifi configurations, I might configure aronia to bridge
traffic between its upstream ethernet port and its wifi card, and to
advertise a separate SSID on its wifi card. As functionality is proven
on aronia, it can be ported to amelanchier.

Manage aronia with Ansible. Let aronia act as a simple, generic server,
and defer the deployment of functionality like newer AP configurations
to a later date.