Rejigger user-ichimonji10 tasks; do private keys
The user-ichimonji10 role has a `main.yml` tasks file which delegates to
other tasks files with syntax like this:

    - include_tasks: servers.yml
      when: "'servers' in group_names"

A downside of this approach is that certain logically atomic tasks
become split across two tasks files. For example, if one wishes to
install `~/.ssh/authorized_keys` on all servers, and uninstall that file
on non-servers, then the following is necessary:

    - include_tasks: servers.yml
      when: "'servers' in group_names"
    - include_tasks: all:!servers.yml
      when: "'servers' not in group_names"

A more straightforward approach is to create one tasks file per topic,
like so:

    - include_tasks: ssh.yml

…and let the tasks file worry about group membership.

Also, this commit updates the user-ichimonji10 role so that it installs
private SSH keys.