Manage nginx SSL files with ansible
Add SSL .crt, .csr and .key files to the vault, in the
`nginx_ssl_domains` variable. Create a new filter, `set_chained_crt`,
that creates a .chained.crt file with Gandi's intermediate SSL
certificate. Modify the nginx role so that it installs the .chained.crt
and .key files in `/etc/nginx/ssl/`.