Cross-reference projects with intersphinx
Create three Python packages:

*   intersphinx-path-root
*   intersphinx-path-1
*   intersphinx-path-2

Create Sphinx-based documentation for each of them. Let the
documentation in intersphinx-path-root serve as the landing page for
all three of them, and let it link to the other plugins.

Create a script that compiles the documentation for all of the Python
packages, collects them into a single directory with a reasonable
structure, and builds them into an nginx-based Docker image.

The build process is unoptimized, and could be sped up through
parallelization. It could also be made more robust, faster to resume
when aborted, and all that. But that's not the point. The point is that
this is a working proof of concept.